Autorización actividades, no pasa el test

+++ app/Http/Controllers/API/ActividadController.php
@@ -6,15 +6,26 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\ActividadResource;
 use App\Models\Actividad;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpKernel\Event\ResponseEvent;
 
-class ActividadController extends Controller
+class ActividadController extends Controller implements HasMiddleware
 {
     public $modelclass = Actividad::class;
 
     /**
      * Display a listing of the resource.
      */
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show']),
+        ];
+    }
+
+
     public function index(Request $request)
     {
         return ActividadResource::collection(
@@ -29,7 +40,9 @@ class ActividadController extends Controller
     public function store(Request $request)
     {
         $actividad = json_decode($request->getContent(), true);
-
+        if (!$request->user()->esAdmin()) {
+            $actividad['user_id'] = $request->user()->id;
+        }
         $actividad = Actividad::create($actividad);
 
         return new ActividadResource($actividad);
@@ -48,6 +61,8 @@ class ActividadController extends Controller
      */
     public function update(Request $request, Actividad $actividad)
     {
+
+        Gate::authorize('update', $actividad);
         $actividadData = json_decode($request->getContent(), true);
 
         $actividad->update($actividadData);
@@ -60,6 +75,7 @@ class ActividadController extends Controller
      */
     public function destroy(Actividad $actividad)
     {
+        Gate::authorize('delete', $actividad);
         try {
             $actividad->delete();
             return response()->json(null, 204);
